cli/pyproject.toml
[project]
name            = 'cli'
version         = '0.1.0'
requires-python = '>=3.9,<3.10'
dependencies    = [
    'toml>=0.10.2'        ,
    'torch>=2.6.0'        ,
    'librosa>=0.11.0'     ,
    'wandb>=0.19.9'       ,
    'python-dotenv>=1.1.0',
    'kirby'
]

[tool.uv.sources]
kirby = { path = '..' }

[project.scripts]
train      = 'cli:train'
inference  = 'cli:inference'
initialize = 'cli:initialize'

[build-system]
requires      = ['hatchling']
build-backend = 'hatchling.build'
